Resolve conflicts:
- ci_test_printers.sh: Keep minimal EVM test (other tests now in pytest)
- call_graph.py: Keep master's graph styling + PR's sorted output
- echidna.py: Keep master's my_super fix for closure capture
- test_printers.py: Keep both PR's snapshot tests and master's specific tests

+++ slither/core/solidity_types/user_defined_type.py
@@ -9,7 +9,7 @@ if TYPE_CHECKING:
     from slither.core.declarations.enum import Enum
     from slither.core.declarations.contract import Contract
 
-# pylint: disable=import-outside-toplevel
+
 class UserDefinedType(Type):
     def __init__(self, t: Union["Enum", "Contract", "Structure"]) -> None:
         from slither.core.declarations.structure import Structure
@@ -37,7 +37,7 @@ class UserDefinedType(Type):
         if isinstance(self._type, Contract):
             return 20, False
         if isinstance(self._type, Enum):
-            return int(math.ceil(math.log2(len(self._type.values)) / 8)), False
+            return math.ceil(math.log2(len(self._type.values)) / 8), False
         if isinstance(self._type, Structure):
             # todo there's some duplicate logic here and slither_core, can we refactor this?
             slot = 0
